+++ Assets/Scripts/WalletModel.cs
@@ -2,20 +2,20 @@ using UnityEngine;
 
 public class WalletModel
 {
-    private decimal _balance;
-    public decimal Balance => _balance;
+    private float _balance;
+    public float Balance => _balance;
 
     // public void Initialize() // [ ] Так вообще можно делать у ScriptableObject? 
     // {
-    //     _balance = decimal.TryParse(balance, out decimal result) ? result : 0m; // [ ] Стоит ли так делать, как я сделал?
+    //     _balance = float.TryParse(balance, out float result) ? result : 0m; // [ ] Стоит ли так делать, как я сделал?
     // }
 
-    public WalletModel(decimal startBalance)
+    public WalletModel(float startBalance)
     {
         _balance = startBalance;
     }
 
-    public bool TrySpend(decimal price)
+    public bool TrySpend(float price)
     {
         if (price > _balance) return false;
